Renamed the fixed function HDF5 data reader that was tuned for Cosmoflow to data_reader_hdf5_legacy.[hpp, cpp].
