+++ include/lbann/data_coordinator/data_coordinator.hpp
@@ -31,7 +31,7 @@
 #include "lbann/utils/dataset.hpp"
 #include "lbann/execution_contexts/execution_context.hpp"
 #ifdef LBANN_HAS_DISTCONV
-#include "lbann/data_readers/data_reader_hdf5.hpp"
+#include "lbann/data_readers/data_reader_hdf5_legacy.hpp"
 #endif // LBANN_HAS_DISTCONV
 
 /** Design docs:
