+++ include/lbann/data_readers/CMakeLists.txt
@@ -5,6 +5,7 @@ set_full_path(THIS_DIR_HEADERS
   data_reader_cifar10.hpp
   data_reader_csv.hpp
   data_reader_image.hpp
+  data_reader_HDF5.hpp
   data_reader_merge_features.hpp
   data_reader_merge_samples.hpp
   data_reader_mnist.hpp
@@ -28,7 +29,7 @@ endif ()
 
 if (LBANN_HAS_DISTCONV)
   list(APPEND THIS_DIR_HEADERS
-    "${CMAKE_CURRENT_SOURCE_DIR}/data_reader_hdf5.hpp")
+    "${CMAKE_CURRENT_SOURCE_DIR}/data_reader_hdf5_legacy.hpp")
 endif ()
 
 if (LBANN_HAS_OPENCV)
